src/pipeline/training_pipeline.py
"""
Training Pipeline Module - Orchestrates the entire training process.
"""
import os
import pandas as pd
import numpy as np
import yaml
import torch
from datetime import datetime
from typing import Dict, List, Optional, Union, Tuple
import logging
import json

from ..data.data_collector import DataCollector
from ..data.data_preprocessor import DataPreprocessor
from ..features.feature_engineering import FeatureEngineer
from ..models.model_runner import ModelTrainer
from ..models.model_evaluator import ModelEvaluator
from ..utils.logger import setup_logger

class TrainingPipeline:
    """
    Orchestrates the end-to-end training process, from data collection to model evaluation.
    """
    
    def __init__(self, config_path: str):
        """
        Initialize the training pipeline.
        
        Args:
            config_path: Path to the configuration file
        """
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
            
        self.logger = setup_logger(__name__)
        self.model_dir = self.config.get('model_dir', 'models')
        os.makedirs(self.model_dir, exist_ok=True)
        
    def run(self):
        """
        Execute the complete training pipeline.
        
        Returns:
            Dictionary with pipeline results including model metrics
        """
        self.logger.info("Starting training pipeline")
        
        # Step 1: Collect data
        self.logger.info("Step 1: Collecting data")
        collector = DataCollector(config=self.config.get('data_collection', {}))
        # Implementation details will go here
        
        # Step 2: Preprocess data
        self.logger.info("Step 2: Preprocessing data")
        preprocessor = DataPreprocessor(config=self.config.get('preprocessing', {}))
        # Implementation details will go here
        
        # Step 3: Engineer features
        self.logger.info("Step 3: Engineering features")
        feature_engineer = FeatureEngineer(config=self.config.get('feature_engineering', {}))
        # Implementation details will go here
        
        # Step 4: Train model
        self.logger.info("Step 4: Training model")
        trainer = ModelTrainer(config=self.config.get('model_training', {}))
        # Implementation details will go here
        
        # Step 5: Evaluate model
        self.logger.info("Step 5: Evaluating model")
        evaluator = ModelEvaluator()
        # Implementation details will go here
        
        self.logger.info("Training pipeline completed successfully")
        return {}
