requirements.txt
# Core libraries
numpy>=1.23.5
pandas>=2.0.1
scikit-learn>=1.2.2

# Deep Learning
torch>=2.0.1

# Data Processing
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.15.0

# API and Serving
fastapi>=0.97.0
uvicorn>=0.22.0
pydantic>=1.10.9

# Monitoring
prometheus-client>=0.17.0

# Statistics
scipy>=1.11.1

# Utilities
tqdm>=4.65.0
joblib>=1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0